[FEATURE] 1) support reading binary file. 2) add fallback for no cookie
1) support reading binary file
Current FileSystem.readFile() API always assumes text file and uses XMLHttpRequest.responseText.
It's required to set XHR.responseType and use XHR.response for reading binary file.
This added additional argument responseType to readFile() and caller can determine the file type.

2) add fallback for the failure of getting cookie
This change adds fallback when the cookie value for Auth host is empty.
The cookie might not be set when a resource is cached and served by Proxy server.
So in that case, I try to get Auth host by window.location value.

And this change also adds a step for getting webidaHost.
Currently, if the cookie is not set, /isloggedin call is requested to 'https://webiga.org'.
So this adds a step to get a host by window.location value.

Change-Id: I35f38d9ab45bbde84bf0601cf33fc9151eeaea93