[BUGFIX] fix a bug that cannot get a service instance scoped as project
An error occurred while getting a service instance scoped as a project because there is no
project scope. So modify code to create a project scope before creating instance of a service.

In addition, it modifies the instance creation part of the service.
Change a second parameter of .apply as an array. It is because must be an array.
And instead of referring to non-existent 'initialized' property to correct to call a 'initialize' function.

And also change the reference to the global self with reference to the local self.
It is because there is no global self.

Signed-off-by: Jugwan Kim <jugwan@webida.org>